+++ src/main/java/com/dh/TPIdiego_cata/repository/IOdontologoRepository.java
@@ -3,5 +3,5 @@ package com.dh.TPIdiego_cata.repository;
 import com.dh.TPIdiego_cata.entity.Odontologo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IOdontologoRepository extends JpaRepository<Odontologo, Long> {
+public interface IOdontologoRepository extends JpaRepository<Odontologo, Integer> {
 }
